add: front gate page(index.js) && rename origin index to blog page
